[FIX] some issues with the display when it comes to custom fonts

+++ platformio/bloated_mp3_player/lib/fonts/include/font_structs.hpp
@@ -13,7 +13,7 @@ struct FontStyleMap {
 };
 
 struct FontHandle {
-    const uint16_t *codes;
+    const uint32_t *codes;
     const uint8_t *bits;
     const uint8_t *widths;
     uint16_t count;
